+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### 0.9.3
 - Updated Spanish translation
 - Check behavior for Foundry 0.9
-- Add "Show Add to Combat Tracker" setting
+- Add "Show Add to Combat Tracker" setting (default: false); if true and Add to CT checkbox is cleared, QE will not add the Actor to the Combat Tracker
 ### 0.9.2
 - Fixed: [#62](https://github.com/spetzel2020/quick-encounters/issues/62) (more improvements planned)
 - Closed: [#53](https://github.com/spetzel2020/quick-encounters/issues/53)